% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.chrTable.R
\name{chrTable.by.4group}
\alias{chrTable.by.4group}
\title{Create Baseline Characteristics Table by three Groups}
\usage{
chrTable.by.4group(
  variable,
  type,
  label,
  data,
  digits,
  summary = "median",
  vargroup,
  vargroup.lab = NULL,
  displaytot = T,
  transpose = F,
  chisqtest = F,
  anova = F,
  useNA = F,
  na = "N/A",
  col1st = ""
)
}
\arguments{
\item{variable}{a vector of variables}

\item{type}{a vector of types either categorical or numerical (\code{"cat"} or \code{"num"})}

\item{label}{a vector of labels shown in the table}

\item{data}{data frame dataset}

\item{digits}{number of decimal points for numerical variable}

\item{summary}{three options are available for the numerical variable to be displayed; If \code{median}, median with min and max is calculated. If \code{median.iq}, median with 25 and 75 percentiles is calculated. If \code{mean}, mean with its standard deviation is calculated. Default is \code{"median"}.}

\item{vargroup}{a variable by which data is staratified}

\item{vargroup.lab}{a vector of grouping labels for \code{vargroup} in the table}

\item{displaytot}{logical; If \code{TRUE}, additional column for the whole cohort is included and default is \code{TRUE}.}

\item{chisqtest}{logical for categorical variable; If \code{TRUE}, Chisquare test is performed. If \code{FALSE}, Fisher's exact test is performed and default is \code{FALSE}.}

\item{anova}{logical for numerical variable; If \code{TRUE}, anova test is performed. If \code{FALSE}, Kruskal-Wallis test is performed and default is \code{FALSE}.}

\item{useNA}{option whether or not \code{NULL} value is displayed for the categorical variable in the table. If \code{TRUE}, character in argument \code{na} is displayed for \code{NULL} value with its percentage. Default is \code{FALSE}.}

\item{na}{character to be displayed for \code{NULL} value if \code{useNA} is \code{TRUE}. Default is \code{"N/A"}.}

\item{col1st}{character label displayed in the 1st cell in the table}
}
\value{
summary matrix with calculated baseline characteristics by group with P values from selected statistical tests
}
\description{
The \code{chrTable.by.4group} function takes data and creates a table for baseline characteristics stratified by 4 groups. Both categorical and numerical variables are included and analyzed together. P values will be bolded in the Markdown document if it is less than 0.05.
}
\examples{
group <- c(rep(1,25), rep(2,25), rep(3,25), rep(4,25))
var1 <- sample(1:100, 100, replace=TRUE)
var2 <- c(sample(1:50, 25, replace=TRUE), sample(25:75, 25, replace=TRUE), sample(30:80, 25, replace=TRUE), sample(50:100, 25, replace=TRUE))
df <- cbind(group, var1, var2)

chrTable.by.4group(variable=c("var1","var2"),type=c("num","num"),
                   label=c("var1", "var2"), data=df, summary="median",
                   vargroup="group", chisqtest=FALSE, anova=FALSE,
                   useNA=FALSE, col1st="")

}
